Change API endpoint structures and formalise ordering filters
