+++ .env.example
@@ -1,5 +1,5 @@
 # Full URL/URI to API. If specified, will disable the Fake API in development environments. Uncomment to connect to live API
-#VUE_APP_API_BASE=http://localhost:8000/api/v1
+#VUE_APP_API_BASE=http://localhost:8000
 
 # Name of the site
 VUE_APP_NAME=UOB Theatre
